refactor: ui -> common
